federated-sovereignty-operator/requirements.txt
kopf==1.35.6
kubernetes==28.1.0
PyYAML==6.0.1
python-dateutil==2.8.2
requests==2.31.0
pydantic==2.5.0
streamlit==1.28.1
folium==0.14.0
pandas==2.1.3
plotly==5.18.0
